+++ Space Shooter/Manager/Factory.cs
@@ -13,19 +13,19 @@ namespace Space_Shooter.Manager
     {
 
         #region Damageable entities
-        public static Game_Player Create_PlayerSpaceship(int x, int y)
+        public static Game_Player Create_PlayerSpaceship(float x, float y)
         {
             return new Game_Player(SpriteManager.Sprites["player1"], x, y);
         }
 
-        public static Bullet_DefaultBullet Create_DefaultBullet(Game_Object owner, int x, int y)
+        public static Bullet_DefaultBullet Create_DefaultBullet(Game_Object owner, float x, float y)
         {
             Bullet_DefaultBullet bullet = new Bullet_DefaultBullet(owner, SpriteManager.Sprites["default_bullet"], x, y);
             GameDataManager.bullets.Add(bullet);
             return bullet;
         }
 
-        public static Enemy_Meteor Create_Meteor(int size, int x, int y, int vx = 0, int vy = 1, int speed = 6, bool AddToData = true)
+        public static Enemy_Meteor Create_Meteor(int size, float x, float y, float vx = 0, float vy = 1, float speed = 6, bool AddToData = true)
         {
             Enemy_Meteor obj = new Enemy_Meteor(SpriteManager.Sprites["meteor_size" + size], x, y, size, vx, vy, speed);
             if (AddToData)
@@ -42,7 +42,7 @@ namespace Space_Shooter.Manager
         ///<summary>
         ///Type is the animation id without explo_ prefix
         ///</summary>
-        public static Game_Animation Create_ani_Explosion(string type, int x, int y)
+        public static Game_Animation Create_ani_Explosion(string type, float x, float y)
         {
             Game_Animation ani = new Game_Animation(SpriteManager.Sprites["explo_" + type], x, y);
             GameDataManager.animations.Add(ani);
@@ -50,7 +50,7 @@ namespace Space_Shooter.Manager
             return ani;
         }
 
-        public static Game_Animation Create_ani_Break(string type, int x, int y)
+        public static Game_Animation Create_ani_Break(string type, float x, float y)
         {
             Game_Animation ani = new Game_Animation(SpriteManager.Sprites["break_" + type], x, y);
             ani.ToCenterPoint(x, y);
